+++ src/App.css
@@ -45,7 +45,8 @@ li {
     position: relative;
     padding: 0 1em;
     opacity: 0.8;
-    transform: translateX(-100vw);
+    transform: translateX(0px);
+    z-index: 2
   }
 
   .sidebar nav {
@@ -107,6 +108,14 @@ a:visited {
   -moz-background-clip: text;
   background-clip: text;
   text-shadow: 0px 3px 3px rgba(255, 255, 255, 0.5);
+}
+
+.ListLinkItem>* {
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  -o-user-select: none;
+  user-select: none;
 }
 
 .current-page {
